requirements.txt
eel==0.16.0
Flask==3.0.3
pyttsx3==2.90
requests==2.31.0
wikipedia==1.4.0
deep-translator==1.11.4
google-genai==0.3.0
python-dotenv==1.0.1
